Load an initial queue as a default if one isn't set
In the rt.cpan.org interface, if you come from a distribution
page, the Report a Bug link will set a queue. But if you click
the Create new ticket button, no queue is set. In RT 5,
some default queue is needed to load the create page, so
pick the first one.